#13 Rename modules (ndtools.* → ndtools.comparison.*)
