+++ ndtools/comparison/__init__.py
@@ -1,0 +1,7 @@
+__all__ = ["collections", "comparable", "operators"]
+
+
+# dependencies
+from . import collections
+from . import comparable
+from . import operators
