+++ ndtools/comparison/collections.py
