+++ ndtools/comparison/comparable.py
