+++ ndtools/comparison/operators.py
